requirements.txt
<<<<<<< HEAD
flask
flask-cors
google-generativeai
gunicorn
=======
flask
flask-cors
google-generativeai
gunicorn
>>>>>>> d673580 (Initial commit)
